requirements.txt
# Data preprocessing and landmark extraction
opencv-python>=4.8.0
mediapipe==0.10.31
numpy>=1.24.0
tqdm>=4.65.0
cvzone>=1.6.1

# Deep learning framework
torch>=2.0.0
torchvision>=0.15.0

# Data handling
pandas>=2.0.0
pyarrow>=12.0.0
Pillow>=9.0.0

# Visualization and metrics
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
tensorboard>=2.13.0

# Optional: for CUDA support
# Install PyTorch with CUDA from: https://pytorch.org/
